Nodes and Elements at the Interface are now copied and get a new id, which isn't taken yet. The crossSectionArea of the Interface Elements is halfed for both parts of the substructure. Any point loads at interface nodes are copied to the new nodes and are halfed for both sides as well. Rest of the code is tidied up and some new functions are implemented to simplify the code.
ATTENTION: constructors of BarElement3d2n and BarElement are changed to cover the case of two inputs. This is needed for the copyElements function. This function instantiates a new BarElement3d2n with only two properties (id and material). This is because if the property nodeArray is set the already created nodes are overwritten and loose attributes gained from copying, like the point loads which are placed on them.
In the BarElement3d2n constructor nodeArray represents the material in the case of only 2 inputs. This should only happen when copy functions calls this constructor.